compiling fixes or something

+++ src/simulation/elements/BOMB.cpp
@@ -33,7 +33,7 @@ int BOMB_update(UPDATE_FUNC_ARGS)
 					pmap[y][x] = 0;
 					for (nxj=-rad; nxj<=rad; nxj++)
 						for (nxi=-rad; nxi<=rad; nxi++)
-							if ((pow(nxi,2))/(pow(rad,2))+(pow(nxj,2))/(pow(rad,2))<=1)
+							if ((pow((float)nxi,2.0f))/(pow((float)rad,2.0f))+(pow((float)nxj,2.0f))/(pow((float)rad,2.0f))<=1)
 
 								if (!(ptypes[pmap[y+nxj][x+nxi]&0xFF].properties&PROP_INDESTRUCTIBLE) && !(ptypes[pmap[y+nxj][x+nxi]&0xFF].properties&PROP_CLONE) && (pmap[y+nxj][x+nxi]&0xFF)!=PT_VIBR) {
 									delete_part(x+nxi, y+nxj, 0);
@@ -47,7 +47,7 @@ int BOMB_update(UPDATE_FUNC_ARGS)
 								}
 					for (nxj=-(rad+1); nxj<=(rad+1); nxj++)
 						for (nxi=-(rad+1); nxi<=(rad+1); nxi++)
-							if ((pow(nxi,2))/(pow((rad+1),2))+(pow(nxj,2))/(pow((rad+1),2))<=1 && !(pmap[y+nxj][x+nxi]&0xFF)) {
+							if ((pow((float)nxi,2.0f))/(pow((float)(rad+1),2.0f))+(pow((float)nxj,2.0f))/(pow((float)(rad+1),2.0f))<=1 && !(pmap[y+nxj][x+nxi]&0xFF)) {
 								nb = create_part(-3, x+nxi, y+nxj, PT_EMBR);
 								if (nb!=-1) {
 									parts[nb].tmp = 0;
